requirements.txt
discord.py
python-dotenv==0.15.0
requests==2.24.0
lxml==4.6.1
asyncpg==0.21.0
